Split client/server bundles.
