+++ package.json
@@ -2,17 +2,21 @@
   "name": "react-easy-universal",
   "version": "1.0.0",
   "description": "Universal Routing & Rendering with React & Redux was too hard. Now it's easy.",
-  "main": "dist/index.js",
+  "main": "dist/",
+  "browser": "dist/",
   "scripts": {
-    "clean": "rimraf dist && rimraf build && mkdir -p dist && mkdir -p build && mkdir -p build/static && cp source/test/fixtures/index.html ./build/index.html",
+    "clean": "rimraf dist && rimraf build && mkdir -p dist && mkdir -p dist/client && mkdir -p build && mkdir -p build/static && cp source/test/fixtures/index.html ./build/index.html",
+    "prestart:client-tests": "npm run clean && npm run build && npm run build:test",
     "start:client-tests": "serve build",
     "start:server": "babel-node source/test/start-server.js",
     "test": "babel-node source/test/index.js",
     "test:e2e": "cross-env NODE_PORT=3000 NODE_PATH=source babel-node source/test/e2e.js",
-    "build:client": "browserify source/test/fixtures/client.js -o build/index.js -t [ babelify --presets [ es2015 react ] ]",
-    "prebuild:test": "npm run -s clean && npm run -s build:client",
-    "build:test": "browserify source/test/client/index.js -o build/static/browser-tests.js -t [ babelify --presets [ es2015 react ] ] && cp build/static/browser-tests.js build/.",
-    "build": "npm run clean && browserify source/index.js -o dist/index.js -t [ babelify --presets [ es2015 react ] ] ",
+    "build:test-client": "browserify -d source/test/fixtures/client.js -o build/index.js -t [ babelify --presets [ es2015 react ] ]",
+    "prebuild:test": "npm run -s clean && npm run -s build:test-client",
+    "build:test": "browserify -d source/test/client/index.js -o build/static/browser-tests.js -t [ babelify --presets [ es2015 react ] ] && cp build/static/browser-tests.js build/.",
+    "build:client": "browserify -d source/client/index.js -o dist/client/index.js -t [ babelify --presets [ es2015 react ] ]",
+    "build:server": "babel -s -d dist/server source/server && babel -s -d dist/shared source/shared",
+    "build": "npm run -s clean && npm run -s build:test-client && npm run -s build:test && npm run -s build:client && npm run -s build:server",
     "prepublish": "npm run build",
     "lint": "eslint .",
     "watch": "watch \"clear && npm run lint -s && npm run test -s\" source",
