+++ source/route-helpers/index.js
@@ -1,0 +1,3 @@
+import { Router, Route } from 'react-router';
+import { connect } from 'react-redux';
+export { Router, Route, connect };
