--- source/test/fixtures/wire-app.js
@@ -1,12 +1,0 @@
-import universal from '../../index';
-
-import routes from './routes';
-import reducers from './reducers';
-
-const wireApp = ({
-  React, app
-}) => universal({
-  React, app, routes, reducers
-});
-
-export default wireApp;
